Next Word Prediction
The next word prediction is a task of predicting the most probable next word in a sequence given some context words. This can be achieved by training a Recurrent Neural Network (RNN) on a large corpus of text data. The RNN is a type of neural network that is capable of processing sequential data by maintaining a hidden state that captures the context of the sequence up to the current time step.